chore(mutator): Tidied up UI attribute features of the Mutator into their own folder. Moved all Mutator out of the SDK space.
