+++ VRCFaceTracking.Core/Params/Data/Mutation/Calibration.cs
@@ -9,7 +9,6 @@ using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
 using VRCFaceTracking.Core.Contracts.Services;
 using VRCFaceTracking.Core.Params.Expressions;
-using VRCFaceTracking.SDK;
 
 namespace VRCFaceTracking.Core.Params.Data.Mutation;
