+++ VRCFaceTracking.Core/Params/Data/Mutation/UI/MutationAttributes.cs
@@ -6,7 +6,7 @@ using System.Reflection;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace VRCFaceTracking.SDK;
+namespace VRCFaceTracking.Core.Params.Data.Mutation;
 
 [AttributeUsage(AttributeTargets.Field, AllowMultiple = false)]
 public class MutationPropertyAttribute : Attribute
